Agents/FlashBrain Orchestrator/requirements.txt
# A2A Protocol Dependencies
a2a-sdk[http-server]>=0.3.16  # Official Google A2A SDK with HTTP server support
httpx>=0.24.0
uvicorn>=0.23.0
click>=8.1.0
pydantic>=2.0.0

# PostgreSQL State Persistence
langgraph-checkpoint-postgres>=1.0.0  # PostgreSQL checkpointer for LangGraph
psycopg[binary]>=3.1.0  # PostgreSQL adapter
psycopg-pool>=3.1.0  # Connection pool for psycopg

# LangChain and LangGraph
langgraph>=1.0.3
langchain-core>=1.0.7
langchain-google-genai>=3.1.0
langchain-google-vertexai>=3.0.3

# Supabase
supabase>=2.24.0

# Other dependencies
python-dotenv>=1.2.1
requests>=2.32.5
